+++ include/session/session_encrypt.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 #include <optional>
 #include <span>
 #include <string>
